Move custom_logger into logging directory
