+++ lib/configuration.js
@@ -23,7 +23,7 @@ convict = require('convict'),
 cjson = require('cjson');
 
 // Side effect - Adds default_bid and dev_bid to express.logger formats
-require('./custom_logger');
+require('./logging/custom_logger');
 
 // verify the proper version of node.js is in use
 try {
