+++ lib/logging/custom_logger.js
